trading_strategy_agent.py
import time
import logging
from swarm import Agent
from cdp import *

# Configure logging
logging.basicConfig(level=logging.DEBUG)
logger = logging.getLogger(__name__)

class TradingStrategyAgent(Agent):
    def __init__(self, name: str, model: str = "gpt-3.5-turbo"):
        # Define instructions first
        instructions = """
        You are a trading strategy agent that develops and executes trading strategies
        on Base. You must document your thought process and maintain clear logs of all
        decisions and actions.
        
        When analyzing market data:
        1. Look for clear patterns and trends
        2. Consider risk management parameters
        3. Propose specific entry and exit points
        4. Document your reasoning
        """
        
        # Initialize the base agent with all parameters
        super().__init__(
            name=name,
            model=model,
            instructions=instructions
        )
        
        logger.info(f"TradingStrategyAgent initialized with model: {model}")
        
    async def develop_strategy(self, market_data):
        """
        Analyzes market data and develops a trading strategy.
        
        Args:
            market_data: JSON string containing market information
        Returns:
            dict: Strategy configuration and analysis
        """
        logger.info(f"Starting strategy development for {self.name}")
        try:
            # Create strategy configuration
            strategy = {
                'name': self.name,
                'type': 'mean_reversion',
                'parameters': {
                    'window_size': 20,
                    'threshold': 2.0
                },
                'description': 'A mean reversion strategy that trades when price deviates significantly from moving average',
                'analysis': {
                    'timestamp': time.time(),
                    'market_conditions': 'Analysis of current market conditions',
                    'rationale': 'Explanation of strategy choice'
                }
            }
            
            logger.info("Strategy development completed successfully")
            return strategy
            
        except Exception as e:
            logger.error(f"Error in strategy development: {str(e)}", exc_info=True)
            raise Exception(f"Strategy development failed: {str(e)}")

    def analyze_market(self, market_data: str) -> str:
        """Analyzes market data and provides insights.
        
        Args:
            market_data: JSON string containing market information
        Returns:
            str: Analysis results and recommendations
        """
        logger.info("Analyzing market data")
        return "Market analysis complete"

    def execute_trade(self, strategy: dict) -> str:
        """Executes a trading strategy.
        
        Args:
            strategy: Dictionary containing strategy parameters
        Returns:
            str: Trade execution results
        """
        logger.info("Executing trade")
        return "Trade executed successfully"
